backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
loguru==0.7.2
openai==1.12.0
cartesia==1.0.3
onnxruntime==1.16.3
supabase==2.3.4
pipecat-ai[daily,cartesia,openai,silero]==0.0.100
gunicorn==21.2.0
